Rag latex fixer/requirements.txt
# Core Dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
torch>=2.0.0

# LaTeX Processing
TexSoup>=0.3.1
pylatexenc>=2.10

# LLM Integration
google-generativeai>=0.3.0
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic-settings>=2.0.0

# Data Processing
pandas>=2.0.0
jsonschema>=4.19.0

# Logging & Monitoring
loguru>=0.7.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
